fix: 워치 앱 크래시 — arm64_32 32비트 Int 오버플로 (Int64 로)
워치(arm64_32)는 Int 가 32비트라 최댓값 약 21억. 이미지 캐시 키가
Int(수정시각 × 1000) 으로 epoch 밀리초(약 1.7조)를 담으려다 Int.max
초과 → 앱 진입(캐릭터 이미지 로드) 즉시 SIGTRAP 크래시.

- CharacterImageStore.activeFileVersion: Int → Int64
- CharacterImageView 애니메이션 프레임 계산도 같은 패턴 → Int64

실기기 재현·심볼라이즈로 원인 확정, 수정 후 재설치하여 정상 실행 확인.

+++ withu/Character/CharacterImageView.swift
@@ -34,7 +34,7 @@ struct CharacterImageView: View {
         #if canImport(UIKit)
         if shouldAnimate {
             TimelineView(.periodic(from: .now, by: 0.7)) { ctx in
-                let frame = Int(ctx.date.timeIntervalSinceReferenceDate / 0.7) % 2
+                let frame = Int(Int64(ctx.date.timeIntervalSinceReferenceDate / 0.7) % 2)
                 singleFrameView(frameIndex: frame)
             }
         } else {
